rename files in bar_styles slash_top to cockpit, and make decorators and constants private
