+++ modules/bar_styles/_constants.py
