+++ modules/bar_styles/silver_blade2.py
@@ -3,7 +3,7 @@ import os
 from libqtile import bar
 from libqtile.config import Screen
 
-from modules.bar_styles.decorators import *
+from modules.bar_styles._decorators import *
 
 
 class SilverBlades2:
