+++ modules/screens.py
@@ -7,14 +7,14 @@ from modules.keys import terminal
 
 from modules.bar_styles.simple_slash_top import SimpleSlashTop
 from modules.bar_styles.simple_slash import SimpleSlash
-from modules.bar_styles.slash_top import SlashTop
+from modules.bar_styles.cockpit import Cockpit
 
 from modules.colors.space_gray import SpaceGray
 from modules.colors.gruvbox import Gruvbox
 
 
 colors = SpaceGray().colors
-screens = SlashTop(colors=colors).screens
+screens = Cockpit(colors=colors).screens
 
 
 # colors = Gruvbox()
